refactor: DeckGenerator 인터페이스로 분리해서 다형성에 열려있는 코드로 사용
