+++ src/main/java/blackjack/domain/card/deck/DeckGenerator.java
@@ -1,0 +1,6 @@
+package blackjack.domain.card.deck;
+
+public interface DeckGenerator {
+
+    Deck generate();
+}
